Start work building a parser library
I intend to use this parser both for parsing the language grammar itself, and for implementing all of TEQL's selectors (including regex selectors). The parser design is not optimized for regular expressions, but I feel like having the regex engine separate from the other selectors the language will support is even less optimal, so I think it should work.

The parser is designed to be non-deterministic--it will simultaneously try all possible branches until the optimal match i found. It is also a streaming or "event-based" parser--it does not load the full text to parse into memory, but works once character at a time, maintaining a list of "marks" (parse events) rather than trying to build an AST while parsing. The marks can be used to construct an AST afterward, if desired, and should also work to implement capture groups.